Custom DC: Allow gcf to handle flex templates
