+++ bigtable_automation/gcf/common.go
@@ -31,47 +31,23 @@ import (
 	"context"
 	"fmt"
 	"log"
-	"path"
 	"strings"
 	"time"
 
 	"cloud.google.com/go/bigtable"
 	"cloud.google.com/go/storage"
 	"github.com/pkg/errors"
-	dataflow "google.golang.org/api/dataflow/v1b3"
 )
 
 const (
-	dataFilePattern    = "cache.csv*"
 	createTableRetries = 3
 	columnFamily       = "csv"
-	// Default region
-	region = "us-central1"
-
-	// NOTE: The following three files represents the state of a BT import. They
-	// get written under:
-	//
-	//		<controlPath>/<TableID>/
-	//
-	// Init: written by borg to start BT import.
-	initFile = "init.txt"
-	// Launched: written by this cloud function to mark launching of BT import job.
-	launchedFile = "launched.txt"
-	// Completed: written by dataflow to mark completion of BT import.
-	completedFile = "completed.txt"
 )
 
 // GCSEvent is the payload of a GCS event.
 type GCSEvent struct {
 	Name   string `json:"name"` // File name in the control folder
 	Bucket string `json:"bucket"`
-}
-
-// joinURL joins url components.
-// path.Join does work well for url, for example gs:// is changaed to gs:/
-func joinURL(base string, paths ...string) string {
-	p := path.Join(paths...)
-	return fmt.Sprintf("%s/%s", strings.TrimRight(base, "/"), strings.TrimLeft(p, "/"))
 }
 
 // parsePath returns the GCS bucket and object from path in the form of gs://<bucket>/<object>
@@ -112,42 +88,6 @@ func writeToGCS(ctx context.Context, path, data string) error {
 	defer w.Close()
 	_, err = fmt.Fprint(w, data)
 	return errors.WithMessagef(err, "Failed to write data to %s/%s", bucket, object)
-}
-
-func launchDataflowJob(
-	ctx context.Context,
-	projectID string,
-	instance string,
-	tableID string,
-	dataPath string,
-	controlPath string,
-	dataflowTemplate string,
-) error {
-	dataflowService, err := dataflow.NewService(ctx)
-	if err != nil {
-		return errors.Wrap(err, "Unable to create dataflow service")
-	}
-	dataFile := joinURL(dataPath, tableID, dataFilePattern)
-	launchedPath := joinURL(controlPath, tableID, launchedFile)
-	completedPath := joinURL(controlPath, tableID, completedFile)
-	params := &dataflow.LaunchTemplateParameters{
-		JobName: tableID,
-		Parameters: map[string]string{
-			"inputFile":          dataFile,
-			"completionFile":     completedPath,
-			"bigtableInstanceId": instance,
-			"bigtableTableId":    tableID,
-			"bigtableProjectId":  projectID,
-			"region":             region,
-		},
-	}
-	log.Printf("[%s/%s] Launching dataflow job: %s -> %s\n", instance, tableID, dataFile, launchedPath)
-	launchCall := dataflow.NewProjectsTemplatesService(dataflowService).Launch(projectID, params)
-	_, err = launchCall.GcsPath(dataflowTemplate).Do()
-	if err != nil {
-		return errors.WithMessagef(err, "Unable to launch dataflow job (%s, %s): %v\n", dataFile, launchedPath)
-	}
-	return nil
 }
 
 func setupBT(ctx context.Context, btProjectID, btInstance, tableID string) error {
