+++ bigtable_automation/gcf/go.mod
@@ -4,8 +4,8 @@ go 1.16
 
 require (
 	cloud.google.com/go/bigtable v1.13.0
-	cloud.google.com/go/storage v1.22.1
+	cloud.google.com/go/storage v1.27.0
 	github.com/GoogleCloudPlatform/functions-framework-go v1.6.1
 	github.com/pkg/errors v0.9.1
-	google.golang.org/api v0.84.0
+	google.golang.org/api v0.105.0
 )
