+++ bigtable_automation/gcf/local/local_flex.yaml
@@ -1,8 +1,9 @@
-projectID: "datcom-website-dev"
-instance: "dc-graph"
-cluster: "dc-graph-c1"
+projectID: "google.com:datcom-store-dev"
+instance: "prophet-test"
+cluster: "prophet-test-c1"
 nodesHigh: "3"
 nodesLow: "1"
 dataflowTemplate: "gs://datcom-dataflow-templates/templates/flex/csv_to_bt.json"
-bucket: "datcom-website-dev-resources"
+dataPath: "gs://prophet_cache"
+controlPath: "gs://automation_control_test"
 tempLocation: "gs://datcom-store-dev-resources/dataflow/tmp"
