+++ bigtable_automation/gcf/prod.go
@@ -72,7 +72,7 @@ func prodInternal(ctx context.Context, e GCSEvent) error {
 	// Get table ID.
 	// e.Name should is like "**/*/branch_2021_01_01_01_01/launched.txt"
 	parts := strings.Split(e.Name, "/")
-	if len(parts) < 3 {
+	if len(parts) < 2 {
 		log.Printf("Ignore irrelevant trigger from file %s", e.Name)
 		return nil
 	}
